+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.1.7-preview.1] - 2019-02-07
+- Now runtime data is readonly. Can not change in inspector.
+- Dialogue Event UnityEvent now support delay by real time.
+
 ## [0.1.6-preview.7] - 2019-02-04
 - Fix end typing event bug.